Backend OSX now builds.

+++ NewGeneBackEnd/Model/Tables/TableInstances/VariableGroup.cpp
@@ -73,7 +73,7 @@ void Table_VG_SET_MEMBER::Load(sqlite3 * db, InputModel * input_model_)
 
 	sqlite3_stmt * stmt = NULL;
 	std::string sql("SELECT * FROM VG_SET_MEMBER");
-	sqlite3_prepare_v2(db, sql.c_str(), sql.size() + 1, &stmt, NULL);
+	sqlite3_prepare_v2(db, sql.c_str(), static_cast<int>(sql.size()) + 1, &stmt, NULL);
 	if (stmt == NULL)
 	{
 		return;
